pyproject.toml
[tool.poetry]
name = "lit-test"
version = "0.1.0"
description = ""
authors = ["Vladislav-76 <sillmaril2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "3.2"
djangorestframework = "3.13"
environs = "^11.0.0"
psycopg2-binary = "^2.9.9"
djoser = "^2.2.2"
drf-yasg = "^1.21.7"
pyotp = "^2.9.0"
gunicorn = "^21.2.0"
celery = "^5.3.6"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
